create form component
